// src/components/shared/checkout-components/form/schemas/checkoutFormSchema.ts
import { z } from "zod";

export const checkoutFromSchema = z.object({
    firstName: z.string().min(2, { message: 'Имя должно содержать не менее 2-х символов' }),
    lastName: z.string().min(2, { message: 'Фамилия должна содержать не менее 2-х символов' }),
    phone: z.string().min(10, { message: 'Введите корректный номер телефона' }),
    email: z.string().email({ message: 'Введите корректную почту' }),
    address: z.string().min(5, { message: 'Введите корректный адрес' }),
    comment: z.string().optional()
})

export type TCheckoutForm = z.infer<typeof checkoutFromSchema>